feat: Add invoice file upload for expenses
Allows users to upload invoice files directly, in addition to providing a URL.

New fields `invoiceData` and `invoiceName` have been added to the `Expense` type to store the base64 encoded file data and the original filename respectively.

The `AddExpenseModal` and `EditExpenseModal` have been updated to support selecting between a URL or an uploaded file for the invoice. The `ExpensesScreen` has been updated to display an "Allegati" column that shows an icon for downloadable attachments.